+++ sigstore/_internal/tuf.py
@@ -20,12 +20,20 @@ from __future__ import annotations
 
 import logging
 import os
+from datetime import datetime, timezone
 from functools import lru_cache
 from pathlib import Path
+from typing import Optional
 from urllib import parse
 
 import appdirs
-from cryptography.x509 import Certificate, load_pem_x509_certificate
+from cryptography.x509 import (
+    Certificate,
+    load_der_x509_certificate,
+    load_pem_x509_certificate,
+)
+from sigstore_protobuf_specs.dev.sigstore.common.v1 import TimeRange
+from sigstore_protobuf_specs.dev.sigstore.trustroot.v1 import TrustedRoot
 from tuf.api import exceptions as TUFExceptions
 from tuf.ngclient import RequestsFetcher, Updater
 
@@ -66,6 +74,28 @@ def _get_dirs(url: str) -> tuple[Path, Path]:
     return (tuf_data_dir / repo_base), (tuf_cache_dir / repo_base)
 
 
+def _is_timerange_valid(period: TimeRange | None, *, allow_expired: bool) -> bool:
+    """
+    Given a `period`, checks that the the current time is not before `start`. If
+    `allow_expired` is `False`, also checks that the current time is not after
+    `end`.
+    """
+    now = datetime.now(timezone.utc)
+
+    # If there was no validity period specified, the key is always valid.
+    if not period:
+        return True
+
+    # Active: if the current time is before the starting period, we are not yet
+    # valid.
+    if now < period.start:
+        return False
+
+    # If we want Expired keys, the key is valid at this point. Otherwise, check
+    # that we are within range.
+    return allow_expired or (period.end is None or now <= period.end)
+
+
 class TrustUpdater:
     """Internal trust root (certificates and keys) downloader.
 
@@ -85,8 +115,6 @@ class TrustUpdater:
         roots, i.e. for the production or staging Sigstore TUF repos.
         """
         self._repo_url = url
-        self._updater: Updater | None = None
-
         self._metadata_dir, self._targets_dir = _get_dirs(url)
 
         # Initialize metadata dir
@@ -125,7 +153,8 @@ class TrustUpdater:
         """
         return cls(STAGING_TUF_URL)
 
-    def _setup(self) -> Updater:
+    @lru_cache()
+    def _updater(self) -> Updater:
         """Initialize and update the toplevel TUF metadata"""
         updater = Updater(
             metadata_dir=str(self._metadata_dir),
@@ -144,14 +173,28 @@ class TrustUpdater:
 
         return updater
 
+    @lru_cache()
+    def _get_trusted_root(self) -> Optional[TrustedRoot]:
+        root_info = self._updater().get_targetinfo("trusted_root.json")
+        if root_info is None:
+            return None
+        path = self._updater().find_cached_target(root_info)
+        if path is None:
+            try:
+                path = self._updater().download_target(root_info)
+            except (
+                TUFExceptions.DownloadError,
+                TUFExceptions.RepositoryError,
+            ) as e:
+                raise TUFError("Failed to download trusted key bundle") from e
+
+        return TrustedRoot().from_json(Path(path).read_bytes())
+
     def _get(self, usage: str, statuses: list[str]) -> list[bytes]:
         """Return all targets with given usage and any of the statuses"""
-        if not self._updater:
-            self._updater = self._setup()
-
         data = []
 
-        targets = self._updater._trusted_set.targets.signed.targets
+        targets = self._updater()._trusted_set.targets.signed.targets
         for target_info in targets.values():
             custom = target_info.unrecognized_fields.get("custom", {}).get("sigstore")
             if (
@@ -159,10 +202,10 @@ class TrustUpdater:
                 and custom.get("status") in statuses
                 and custom.get("usage") == usage
             ):
-                path = self._updater.find_cached_target(target_info)
+                path = self._updater().find_cached_target(target_info)
                 if path is None:
                     try:
-                        path = self._updater.download_target(target_info)
+                        path = self._updater().download_target(target_info)
                     except (
                         TUFExceptions.DownloadError,
                         TUFExceptions.RepositoryError,
@@ -187,7 +230,21 @@ class TrustUpdater:
 
         May download files from the remote repository.
         """
-        ctfes = self._get("CTFE", ["Active"])
+        ctfes = []
+
+        trusted_root = self._get_trusted_root()
+        if trusted_root:
+            for key in trusted_root.ctlogs:
+                if not _is_timerange_valid(
+                    key.public_key.valid_for, allow_expired=False
+                ):
+                    continue
+                key_bytes = key.public_key.raw_bytes
+                if key_bytes:
+                    ctfes.append(key_bytes)
+        else:
+            ctfes = self._get("CTFE", ["Active"])
+
         if not ctfes:
             raise MetadataError("CTFE keys not found in TUF metadata")
         return ctfes
@@ -197,7 +254,21 @@ class TrustUpdater:
 
         May download files from the remote repository.
         """
-        keys = self._get("Rekor", ["Active"])
+        keys = []
+
+        trusted_root = self._get_trusted_root()
+        if trusted_root:
+            for key in trusted_root.tlogs:
+                if not _is_timerange_valid(
+                    key.public_key.valid_for, allow_expired=False
+                ):
+                    continue
+                key_bytes = key.public_key.raw_bytes
+                if key_bytes:
+                    keys.append(key_bytes)
+        else:
+            keys = self._get("Rekor", ["Active"])
+
         if len(keys) != 1:
             raise MetadataError("Did not find one active Rekor key in TUF metadata")
         return keys
@@ -207,9 +278,29 @@ class TrustUpdater:
 
         May download files from the remote repository.
         """
+        certs = []
+
+        trusted_root = self._get_trusted_root()
         # Return expired certificates too: they are expired now but may have
         # been active when the certificate was used to sign.
-        certs = self._get("Fulcio", ["Active", "Expired"])
+        if trusted_root:
+            for ca in trusted_root.certificate_authorities:
+                if not _is_timerange_valid(ca.valid_for, allow_expired=True):
+                    continue
+                certs.extend(
+                    [
+                        load_der_x509_certificate(cert.raw_bytes)
+                        for cert in ca.cert_chain.certificates
+                    ]
+                )
+        else:
+            certs = [
+                load_pem_x509_certificate(c)
+                for c in self._get(
+                    "Fulcio",
+                    ["Active", "Expired"],
+                )
+            ]
         if not certs:
             raise MetadataError("Fulcio certificates not found in TUF metadata")
-        return [load_pem_x509_certificate(c) for c in certs]
+        return certs